+++ docs/changelog.md
@@ -49,6 +49,7 @@ As features stabilize some brief notes about them will accumulate here.
 * Unix: Clicking a URL when no browser is open could cause wezterm to hang until the newly opened browser is closed. [#2245](https://github.com/wez/wezterm/issues/2245)
 * Quickselect: now selects the bottom-most match rather than the top-most match. [#2250](https://github.com/wez/wezterm/issues/2250)
 * Mux: `wezterm.mux.set_active_workspace` didn't update the current window to match the newly activated workspace. [#2248](https://github.com/wez/wezterm/issues/2248)
+* Overlays such as debug and launcher menu now handle resize better
 
 #### Updated
 * Bundled harfbuzz to 4.4.1